Moved Spline from path to util.
