+++ aikido/src/path/CMakeLists.txt
@@ -4,5 +4,6 @@ aikido_add_library("${PROJECT_NAME}_path"
 )
 target_link_libraries("${PROJECT_NAME}_path"
   "${PROJECT_NAME}_statespace"
+  "${PROJECT_NAME}_util"
   ${DART_LIBRARIES}
 )
